+++ apps/docs/components/Header.tsx
@@ -3,10 +3,7 @@ import Link from 'next/link'
 import { Logo } from '@compai/logo'
 import pkg from '../../../packages/gui/package.json'
 
-type HeaderProps = {
-  showDocsLink?: boolean
-}
-export const Header = ({ showDocsLink }: HeaderProps) => {
+export const Header = () => {
   const id = useId()
 
   return (
@@ -41,7 +38,7 @@ export const Header = ({ showDocsLink }: HeaderProps) => {
           <Logo height={20} width={20} seed={id} />
         </span>
       </a>
-      <Link href="/home" passHref={true}>
+      <Link href="/" passHref={true}>
         <a
           sx={{
             fontWeight: 500,
@@ -59,56 +56,53 @@ export const Header = ({ showDocsLink }: HeaderProps) => {
           </span>
         </a>
       </Link>
-      <div sx={{ ml: 'auto', display: 'flex' }}>
-        {showDocsLink ? (
-          <Link href="/home" passHref={true}>
-            <a
-              sx={{
-                fontWeight: 500,
-                color: 'text',
-                textDecoration: 'none',
-                display: 'flex',
-                alignItems: 'center',
-                fontSize: [0, 1, 2],
-                px: 3,
-              }}
-            >
-              Docs
-            </a>
-          </Link>
-        ) : null}
-        <Link href="/properties" passHref={true}>
-          <a
-            sx={{
-              fontWeight: 500,
-              color: 'text',
-              textDecoration: 'none',
-              display: 'flex',
-              alignItems: 'center',
-              fontSize: [0, 1, 2],
-              px: 3,
-            }}
-          >
-            Properties
-          </a>
-        </Link>
+      <Link href="/home" passHref={true}>
         <a
           sx={{
+            fontWeight: 500,
+            color: 'text',
             textDecoration: 'none',
-            color: 'currentColor',
-            fontWeight: 600,
-            transition: 'color .2s ease-in-out',
+            display: 'flex',
+            alignItems: 'center',
+            fontSize: [0, 1, 2],
             px: 3,
-            fontSize: [0, 1, 1],
-            ':hover': {
-              color: 'primary',
-            },
+            ml: 'auto',
           }}
-          href="https://github.com/components-ai/css.gui"
         >
-          GitHub
+          Docs
         </a>
-      </div>
+      </Link>
+      <Link href="/properties" passHref={true}>
+        <a
+          sx={{
+            fontWeight: 500,
+            color: 'text',
+            textDecoration: 'none',
+            display: 'flex',
+            alignItems: 'center',
+            fontSize: [0, 1, 2],
+            px: 3,
+          }}
+        >
+          Properties
+        </a>
+      </Link>
+      <a
+        sx={{
+          textDecoration: 'none',
+          color: 'currentColor',
+          fontWeight: 600,
+          transition: 'color .2s ease-in-out',
+          px: 3,
+          fontSize: [0, 1, 1],
+          ':hover': {
+            color: 'primary',
+          },
+        }}
+        href="https://github.com/components-ai/css.gui"
+      >
+        GitHub
+      </a>
     </header>
   )
 }